% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kruskal_wallis.R
\name{kruswall}
\alias{kruswall}
\title{Prueba de Kruskal-Wallis}
\usage{
kruswall(y, alpha)
}
\arguments{
\item{y}{(list) datos}

\item{alpha}{(numeric) nivel de significancia}
}
\value{
una lista con el número de grupos, la longitud de cada grupo, número total de observaciones, media del rango por grupo, media del rango total, el estadístico de prueba, los grados de libertad, el valor crítico y el p-value.
}
\description{
Realiza la prueba de Kruskal-Wallis, es decir,
determina si hay diferencias significativas
entre las medianas de tres o más grupos independientes
cuando no se cumplen los supuestos de los métodos paramétricos.
}
\examples{
  \dontrun{
#\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%\%
#Ejemplo 1:
#El conjunto de datos “PlantGrowth” incorporado en el lenguaje de programación R contiene dos variables de interes:
#weight: Es una variable numérica que representa el peso del crecimiento de las plantas. Es la variable de interés principal en este conjunto de datos.
#group: Es una variable categórica que indica el grupo de tratamiento al que pertenece cada planta. Hay tres grupos de tratamiento: “ctrl”, “trt1” y “trt2” y cada grupo contiene 10 observaciones.
#El objetivo de utilizar la prueba de Kruskal-Wallis con el conjunto de datos “PlantGrowth” sería determinar si existen diferencias significativas en el crecimiento de las plantas entre los diferentes grupos de tratamiento (“ctrl”, “trt1” y “trt2”).

#install.packages("datasets")
#data("PlantGrowth")
#PlantGrowth
datos = list(
 ctrl<-c(4.17,5.58,5.18,6.11, 4.50,4.61,5.17,4.53, 5.33,5.14),
 trt1<-c(4.81,4.17,4.41,3.59,5.87,3.83,6.03, 4.89,4.32,4.69),
 trt2<-c(6.31, 5.12,5.54,5.50, 5.37,5.29,4.92,6.15, 5.80, 5.26 )
)
#Cargar la libreria
library("KruskalWalis")
kruswall(y=datos,alpha=alpha)
  }
}
